move tests to be closer to file tested
